Move clutter into helper.ts from MaterialSlider.tsx
