+++ components/paragraphs/MaterialSlider/MaterialSlider.tsx
@@ -2,10 +2,11 @@
 
 import { useWindowSize } from "@uidotdev/usehooks"
 import "keen-slider/keen-slider.min.css"
-import { KeenSliderOptions, KeenSliderPlugin, useKeenSlider } from "keen-slider/react"
+import { useKeenSlider } from "keen-slider/react"
 import Link from "next/link"
 import React, { useEffect, useState } from "react"
 
+import { WheelControls, defaultSliderOptions } from "@/components/paragraphs/MaterialSlider/helper"
 import { Button } from "@/components/shared/button/Button"
 import Icon from "@/components/shared/icon/Icon"
 import WorkCard, { WorkCardSkeleton } from "@/components/shared/workCard/WorkCard"
@@ -21,96 +22,6 @@ import { resolveUrl } from "@/lib/helpers/helper.routes"
 type MaterialSliderProps = {
   works: ComplexSearchForWorkTeaserQuery["complexSearch"]["works"] | undefined
   title: ParagraphGoMaterialSliderAutomatic["title"] | ParagraphGoMaterialSliderManual["title"]
-}
-
-const defaultSliderOptions: KeenSliderOptions = {
-  initial: 0,
-  slides: {
-    origin: "auto",
-    spacing: 12,
-    perView: 1.1,
-  },
-  breakpoints: {
-    "(min-width: 768px)": {
-      slides: {
-        origin: "auto",
-        spacing: 12,
-        perView: () => {
-          return 2.1
-        },
-      },
-    },
-    "(min-width: 1024px)": {
-      slides: {
-        spacing: 24,
-        perView: () => {
-          return 3
-        },
-      },
-    },
-  },
-}
-
-const WheelControls: KeenSliderPlugin = slider => {
-  let touchTimeout: ReturnType<typeof setTimeout>
-  let position: {
-    x: number
-    y: number
-  }
-  let wheelActive: boolean
-
-  function dispatch(e: WheelEvent, name: string) {
-    position.x -= e.deltaX
-    position.y -= e.deltaY
-    slider.container.dispatchEvent(
-      new CustomEvent(name, {
-        detail: {
-          x: position.x,
-          y: position.y,
-        },
-      })
-    )
-  }
-
-  function wheelStart(e: WheelEvent) {
-    position = {
-      x: e.pageX,
-      y: e.pageY,
-    }
-    dispatch(e, "ksDragStart")
-  }
-
-  function wheel(e: WheelEvent) {
-    dispatch(e, "ksDrag")
-  }
-
-  function wheelEnd(e: WheelEvent) {
-    dispatch(e, "ksDragEnd")
-  }
-
-  function eventWheel(e: WheelEvent) {
-    if (e.deltaX === 0) {
-      return
-    }
-
-    e.preventDefault()
-    if (!wheelActive) {
-      wheelStart(e)
-      wheelActive = true
-    }
-    wheel(e)
-    clearTimeout(touchTimeout)
-    touchTimeout = setTimeout(() => {
-      wheelActive = false
-      wheelEnd(e)
-    }, 50)
-  }
-
-  slider.on("created", () => {
-    slider.container.addEventListener("wheel", eventWheel, {
-      passive: false,
-    })
-  })
 }
 
 const MaterialSlider = ({ works, title }: MaterialSliderProps) => {
