+++ components/paragraphs/MaterialSlider/helper.ts
@@ -1,0 +1,90 @@
+import { KeenSliderOptions, KeenSliderPlugin } from "keen-slider/react"
+
+export const defaultSliderOptions: KeenSliderOptions = {
+  initial: 0,
+  slides: {
+    origin: "auto",
+    spacing: 12,
+    perView: 1.1,
+  },
+  breakpoints: {
+    "(min-width: 768px)": {
+      slides: {
+        origin: "auto",
+        spacing: 12,
+        perView: () => {
+          return 2.1
+        },
+      },
+    },
+    "(min-width: 1024px)": {
+      slides: {
+        spacing: 24,
+        perView: () => {
+          return 3
+        },
+      },
+    },
+  },
+}
+
+export const WheelControls: KeenSliderPlugin = slider => {
+  let touchTimeout: ReturnType<typeof setTimeout>
+  let position: {
+    x: number
+    y: number
+  }
+  let wheelActive: boolean
+
+  function dispatch(e: WheelEvent, name: string) {
+    position.x -= e.deltaX
+    position.y -= e.deltaY
+    slider.container.dispatchEvent(
+      new CustomEvent(name, {
+        detail: {
+          x: position.x,
+          y: position.y,
+        },
+      })
+    )
+  }
+  function wheelStart(e: WheelEvent) {
+    position = {
+      x: e.pageX,
+      y: e.pageY,
+    }
+    dispatch(e, "ksDragStart")
+  }
+
+  function wheel(e: WheelEvent) {
+    dispatch(e, "ksDrag")
+  }
+
+  function wheelEnd(e: WheelEvent) {
+    dispatch(e, "ksDragEnd")
+  }
+
+  function eventWheel(e: WheelEvent) {
+    if (e.deltaX === 0) {
+      return
+    }
+
+    e.preventDefault()
+    if (!wheelActive) {
+      wheelStart(e)
+      wheelActive = true
+    }
+    wheel(e)
+    clearTimeout(touchTimeout)
+    touchTimeout = setTimeout(() => {
+      wheelActive = false
+      wheelEnd(e)
+    }, 50)
+  }
+
+  slider.on("created", () => {
+    slider.container.addEventListener("wheel", eventWheel, {
+      passive: false,
+    })
+  })
+}
